+++ api/python/arrus/devices/us4r.py
@@ -44,18 +44,20 @@ class Backplane:
     """
     Digital backplane of the us4R device.
     """
+    def __init__(self, us4r):
+        self._us4r = us4r
 
     def get_serial_number(self) -> str:
         """
         Returns serial number of the digital backplane.
         """
-        return ""
+        return self._us4r._handle.getBackplaneSerialNumber()
 
     def get_revision(self) -> str:
         """
         Returns revision number of the digital backplane.
         """
-        return ""
+        return self._us4r._handle.getBackplaneRevision()
 
 
 class Us4R(Device, Ultrasound):
@@ -72,7 +74,7 @@ class Us4R(Device, Ultrasound):
                                    self._handle.getDeviceId().getOrdinal())
         # Context for the currently running sequence.
         self._kernel_context = None
-        self._backplane = Backplane()
+        self._backplane = Backplane(self)
 
     def get_device_id(self):
         return self._device_id
@@ -211,7 +213,7 @@ class Us4R(Device, Ultrasound):
 
     def set_test_pattern(self, pattern):
         """
-        Sets given test ADC test patter to be run by Us4OEM components.
+        Sets given test ADC test pattern to be run by Us4OEM components.
         """
         test_pattern_core = arrus.utils.core.convert_to_test_pattern(pattern)
         self._handle.setTestPattern(test_pattern_core)
@@ -281,6 +283,7 @@ class Us4R(Device, Ultrasound):
         :param addr: register address (8-bit)
         """
         return self._handle.getAfe(addr)
+
 
     @property
     def channels_mask(self):